Refactor common code into a base class
